Moved buttons to common components
