+++ frontend/src/components/common/buttons/NavigateButton.tsx
@@ -1,12 +1,12 @@
-const OpenHistoryButton = (props: { onClick: () => void }) => {
+const NavigateButton = (props: { navigate: () => void; name: string }) => {
   return (
     <button
       className="bg-blue-600 hover:bg-blue-300 pt-2 pb-2 pl-4 pr-4 border-gray-300  border-2 text-white"
-      onClick={props.onClick}
+      onClick={props.navigate}
     >
-      Open History
+      {props.name}
     </button>
   );
 };
 
-export default OpenHistoryButton;
+export default NavigateButton;
