+++ frontend/src/components/common/buttons/OpenButton.tsx
@@ -1,12 +1,12 @@
-const CreateTaskButton = (props: { onClick: () => void }) => {
+const OpenButton = (props: { open: () => void; name: string }) => {
   return (
     <button
       className="bg-green-600 hover:bg-green-300 pt-2 pb-2 pl-4 pr-4 border-gray-300  border-2 text-white"
-      onClick={props.onClick}
+      onClick={props.open}
     >
-      Create Task
+      {props.name}
     </button>
   );
 };
 
-export default CreateTaskButton;
+export default OpenButton;
